Change scoring system. Scores are now integers since decimals seemed unpolished. Starter complexities are no longer scored in the range of 100s, but rather 10s of skill points.

+++ internal/database/entities.go
@@ -111,7 +111,7 @@ type User struct {
 	DateRegistered      time.Time         `gorm:"not null" json:"-"`
 	PasswordResetToken  WrappedNullString `gorm:"type:char(128);null;default:null" json:"-"`
 	PasswordResetExpiry WrappedNullTime   `gorm:"null;default:null" json:"-"`
-	TotalScore          float32           `gorm:"not null;default:0" json:"totalScore"`
+	TotalScore          int               `gorm:"not null;default:0" json:"totalScore"`
 }
 
 type RefreshToken struct {
